// src/app/client-app/Features/Tour-Guides/tour-guides/interfaces/tour-guide.ts
export interface TourGuide {
  guideID: string;
  guideName: string;
  licenseNumber: string;
  languages: string;
  areasCovered: string;
  pricePerHour: number;
  contactEmail: string;
  contactPhone: string;
  averageRating: number;
  totalBookings: number;
  isAvailable: boolean;
  photoUrls: string[];
}